requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
sqlalchemy==2.0.35
alembic==1.13.3
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.9.2
pydantic-settings==2.5.2
mercadopago==2.3.0
httpx==0.27.2
python-dotenv==1.0.1
pillow==10.4.0
openpyxl==3.1.5
reportlab==4.2.5
pytest==8.3.3
pytest-asyncio==0.23.8
httpx==0.27.2
